Initial Commit: A Very Basic Chicken Killin' Script
This is the first commit for a chicken killer script.  My end goal is to
support both chicken locations in lumbridge and have some pathing stuff
to make starting the bot easier.  The purpose of this script is to ready
accounts for farming via combat.

The features of this bot are:
Support for the big patch of chickens (west of river Lumb).
Returns to center of pasture if no chickens are found.
Kills chickens.